+++ test/features/distribute-assert/packages/api/package.json
@@ -34,6 +34,6 @@
   },
   "dependencies": {
     "@nestia/fetcher": "^1.4.0",
-    "typia": "^4.1.4"
+    "typia": "^4.1.5"
   }
 }
