+++ test/features/simulate/src/api/functional/bbs/articles/index.ts
@@ -220,6 +220,71 @@ export namespace at {
 }
 
 /**
+ * Get first article of a day.
+ * 
+ * @param section Section code
+ * @param date Target data
+ * @returns The first article info
+ * 
+ * @controller BbsArticlesController.first()
+ * @path GET /bbs/:section/articles/first/:date
+ * @nestia Generated by Nestia - https://github.com/samchon/nestia
+ */
+export async function first(
+    connection: IConnection,
+    section: string,
+    date: string,
+): Promise<first.Output> {
+    return !!connection.simulate
+        ? first.simulate(
+              connection,
+              section,
+              date,
+          )
+        : Fetcher.fetch(
+              connection,
+              first.ENCRYPTED,
+              first.METHOD,
+              first.path(section, date),
+          );
+}
+export namespace first {
+    export type Output = Primitive<IBbsArticle>;
+
+    export const METHOD = "GET" as const;
+    export const PATH: string = "/bbs/:section/articles/first/:date";
+    export const ENCRYPTED: Fetcher.IEncrypted = {
+        request: false,
+        response: false,
+    };
+
+    export const path = (section: string, date: string): string => {
+        return `/bbs/${encodeURIComponent(section ?? "null")}/articles/first/${encodeURIComponent(date ?? "null")}`;
+    }
+    export const random = (g?: Partial<typia.IRandomGenerator>): Output =>
+        typia.random<Output>(g);
+    export const simulate = async (
+        connection: IConnection,
+        section: string,
+        date: string,
+    ): Promise<Output> => {
+        const assert = NestiaSimulator.assert({
+            method: METHOD,
+            host: connection.host,
+            path: path(section, date)
+        });
+        assert.param("section")("string")(() => typia.assert(section));
+        assert.param("date")("date")(() => typia.assert(date));
+        return random(
+            typeof connection.simulate === 'object' &&
+                connection.simulate !== null
+                ? connection.simulate
+                : undefined
+        );
+    }
+}
+
+/**
  * Store a new article.
  * 
  * @param section Section code
@@ -349,71 +414,6 @@ export namespace update {
         assert.param("section")("string")(() => typia.assert(section));
         assert.param("id")("uuid")(() => typia.assert(id));
         assert.body(() => typia.assert(input));
-        return random(
-            typeof connection.simulate === 'object' &&
-                connection.simulate !== null
-                ? connection.simulate
-                : undefined
-        );
-    }
-}
-
-/**
- * Get first article of a day.
- * 
- * @param section Section code
- * @param date Target data
- * @returns The first article info
- * 
- * @controller BbsArticlesController.first()
- * @path GET /bbs/:section/articles/first/:date
- * @nestia Generated by Nestia - https://github.com/samchon/nestia
- */
-export async function first(
-    connection: IConnection,
-    section: string,
-    date: string,
-): Promise<first.Output> {
-    return !!connection.simulate
-        ? first.simulate(
-              connection,
-              section,
-              date,
-          )
-        : Fetcher.fetch(
-              connection,
-              first.ENCRYPTED,
-              first.METHOD,
-              first.path(section, date),
-          );
-}
-export namespace first {
-    export type Output = Primitive<IBbsArticle>;
-
-    export const METHOD = "GET" as const;
-    export const PATH: string = "/bbs/:section/articles/first/:date";
-    export const ENCRYPTED: Fetcher.IEncrypted = {
-        request: false,
-        response: false,
-    };
-
-    export const path = (section: string, date: string): string => {
-        return `/bbs/${encodeURIComponent(section ?? "null")}/articles/first/${encodeURIComponent(date ?? "null")}`;
-    }
-    export const random = (g?: Partial<typia.IRandomGenerator>): Output =>
-        typia.random<Output>(g);
-    export const simulate = async (
-        connection: IConnection,
-        section: string,
-        date: string,
-    ): Promise<Output> => {
-        const assert = NestiaSimulator.assert({
-            method: METHOD,
-            host: connection.host,
-            path: path(section, date)
-        });
-        assert.param("section")("string")(() => typia.assert(section));
-        assert.param("date")("date")(() => typia.assert(date));
         return random(
             typeof connection.simulate === 'object' &&
                 connection.simulate !== null
