--- test/features/variable/src/api/functional/variable/$try/index.ts
@@ -1,37 +1,0 @@
-/**
- * @packageDocumentation
- * @module api.functional.variable.$try
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
- */
-//================================================================
-import { Fetcher } from "@nestia/fetcher";
-import type { IConnection } from "@nestia/fetcher";
-
-/**
- * @controller VariableController.try()
- * @path GET /variable/try
- * @nestia Generated by Nestia - https://github.com/samchon/nestia
- */
-export async function $try(
-    connection: IConnection,
-): Promise<void> {
-    return Fetcher.fetch(
-        connection,
-        $try.ENCRYPTED,
-        $try.METHOD,
-        $try.path(),
-    );
-}
-export namespace $try {
-
-    export const METHOD = "GET" as const;
-    export const PATH: string = "/variable/try";
-    export const ENCRYPTED: Fetcher.IEncrypted = {
-        request: false,
-        response: false,
-    };
-
-    export const path = (): string => {
-        return `/variable/try`;
-    }
-}
