+++ test/features/variable/src/test/features/api/automated/test_api_variable_$catch.ts
@@ -1,9 +1,9 @@
 import api from "./../../../../api";
 
-export const test_api_variable_$try_$try = async (
+export const test_api_variable_$catch = async (
     connection: api.IConnection
 ): Promise<void> => {
-    await api.functional.variable.$try.$try(
+    await api.functional.variable.$catch(
         connection,
     );
 };
